relocate password field from Input component
