+++ client/src/shared/components/FormElements/Password.css
@@ -1,0 +1,9 @@
+.form-input span {
+	cursor: pointer;
+	border: 1px solid #ccc;
+	border-left: none;
+	background: #ebebeb;
+}
+.form-input span:active {
+	background: #979797;
+}
